+++ History.md
@@ -10,6 +10,7 @@ Unreleased
 * added; `invoice.InvoiceNumberPrefix` and `invoice.InvoiceNumberWithPrefix()`
 * added; `transaction.GetInvoice()`
 * added; `invoice.GetOriginalInvoice()`
+* added; CreditCardType.DinersClub
 
 1.1.9 (stable) / 2015-04-01
 ==================